Make randomizer fair
In tetris it isn't considered to be good to
randomize just every block. There should be
certain repeatability enforced in randomess.

This change should be probably best made in
the tetris_core lib itself, however I made
it quickly here now since the Randomizer
is only ever used to get that one specific
value to decide on the next block and lets
me still use the official library (unforked).