Use Instant instead of SystemTime
This fixes times where the delay was either too
little or some other kind of drift/adjustment
happend which cause a negative duration and a in
some cases still unhandled error.

Instant is the correct thing to use in this cases
anyway and it will fix some rare crashes during
the game.